Basic training and evaluation pipeline complete
